requirements.txt
matplotlib==3.1.3
pydub==0.23.1
numpy==1.18.1
librosa==0.7.2
ImageHash==4.0
Pillow==7.1.2
PyQt5==5.14.2
scikit_learn==0.22.2.post1
ffmpeg 
ffprobe
